refactor: User 관련 DTO 패키지 분리
