+++ src/main/java/woowacrew/user/service/UserInternalService.java
@@ -3,6 +3,7 @@ package woowacrew.user.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacrew.user.domain.*;
+import woowacrew.user.dto.UserApproveDto;
 import woowacrew.user.service.exception.NotExistUserException;
 
 import java.time.LocalDate;
